+++ CHANGES.rst
@@ -16,12 +16,15 @@ assign_wcs
 - Added a wavelength correction for the effective velocity of JWST
   relative to the barycenter.                                  [#2359, #2406]
 
+- NRC_TSGRISM assigns source location to set pixel [#1235]
 associations
 ------------
 
 - Implemented Rule for Level 2 Nirspec Fixed Slit background. [#2307]
 - Handle both numeric and named slits for Level3 products. [#2330]
 - Remove MIR_LRS-SLITLESS and NIS_SOSS from the permanent TSO list. [#2330]
+- Implement new Level2a rule `Asn_Lv2NRSLAMP`. [#2177]
+- Allow "N/A" as a valid, but False, value in association pools. [#2334]
 
 background
 ----------
@@ -62,6 +65,8 @@ datamodels
 
 - New info method, similar to the method in astropy fits [#2268]
 
+- Removed BaseExtension class, it was not being used [#2003]
+
 dq_init
 -------
 
@@ -79,6 +84,7 @@ extract_1d
 
 extract_2d
 ----------
+- NRC_TSGRISM implemented with set source location and extraction options [#1710, #1235]
 
 firstframe
 ----------
@@ -152,6 +158,10 @@ pipeline
   step from running. [#2318]
 
 - Enable resample_spec to run on MIRI fixed slit data in calspec2 [#2424]
+
+- Implement new `Spec2Pipeline` configuration for NIRSpec LAMP exposures [#2174]
+
+- Implement specific exit status for "no science on detector" [#2336]
 
 ramp_fitting
 ------------